UserUtils has been integrated into Users
